컴포넌트 이름 변경 등 정리
