+++ src/pages/shelves/detail/SearchModal/index.jsx
@@ -3,77 +3,25 @@ import { css, jsx } from '@emotion/core';
 import React from 'react';
 import { connect } from 'react-redux';
 
-import BottomActionBar from '../../../components/BottomActionBar';
-import { ButtonType } from '../../../components/ActionBar/constants';
-import { ACTION_BAR_HEIGHT } from '../../../components/ActionBar/styles';
-import Empty from '../../../components/Empty';
-import FixedToolbarView from '../../../components/FixedToolbarView';
-import FlexBar from '../../../components/FlexBar';
-import { ResponsivePaginatorWithHandler } from '../../../components/ResponsivePaginator';
-import SearchBox from '../../../components/SearchBox';
-import SkeletonBooks from '../../../components/Skeleton/SkeletonBooks';
-import { LIBRARY_ITEMS_LIMIT_PER_PAGE } from '../../../constants/page';
-import ViewType from '../../../constants/viewType';
-import * as searchRequests from '../../../services/purchased/search/requests';
-import * as selectionActions from '../../../services/selection/actions';
-import * as selectionSelectors from '../../../services/selection/selectors';
-import * as shelfSelectors from '../../../services/shelf/selectors';
-import ArrowLeft from '../../../svgs/ArrowLeft.svg';
-import SearchIcon from '../../../svgs/Search.svg';
-import * as paginationUtils from '../../../utils/pagination';
-import Responsive, { ResponsiveBooks } from '../../base/Responsive';
+import BottomActionBar from '../../../../components/BottomActionBar';
+import { ButtonType } from '../../../../components/ActionBar/constants';
+import { ACTION_BAR_HEIGHT } from '../../../../components/ActionBar/styles';
+import Empty from '../../../../components/Empty';
+import FixedToolbarView from '../../../../components/FixedToolbarView';
+import { ResponsivePaginatorWithHandler } from '../../../../components/ResponsivePaginator';
+import SkeletonBooks from '../../../../components/Skeleton/SkeletonBooks';
+import { LIBRARY_ITEMS_LIMIT_PER_PAGE } from '../../../../constants/page';
+import ViewType from '../../../../constants/viewType';
+import * as searchRequests from '../../../../services/purchased/search/requests';
+import * as selectionActions from '../../../../services/selection/actions';
+import * as selectionSelectors from '../../../../services/selection/selectors';
+import * as shelfSelectors from '../../../../services/shelf/selectors';
+import SearchIcon from '../../../../svgs/Search.svg';
+import * as paginationUtils from '../../../../utils/pagination';
+import { ResponsiveBooks } from '../../../base/Responsive';
+import NavigationBar from './NavigationBar';
+import SearchBar from './SearchBar';
 import SearchBooks from './SearchBooks';
-
-const navigationBarStyles = {
-  wrapper: css`
-    background-color: #0077d9;
-    width: 100%;
-  `,
-  bar: css`
-    height: 46px;
-    display: flex;
-    align-items: center;
-  `,
-  iconWrapper: css`
-    display: block;
-    padding: 13px 10px 13px 0;
-  `,
-  icon: css`
-    width: 16px;
-    height: 16px;
-    fill: white;
-  `,
-  title: css`
-    font-size: 16px;
-    font-weight: bold;
-    color: white;
-    height: 30px;
-    line-height: 30px;
-  `,
-};
-
-const searchBarStyles = {
-  bar: css`
-    border-bottom: 1px solid #d1d5d9;
-    box-shadow: 0 1px 6px 0 rgba(0, 0, 0, 0.04);
-    background-color: #f3f4f5;
-  `,
-  cancelButton: css`
-    display: block;
-    margin-left: 14px;
-    border-radius: 4px;
-    box-shadow: 1px 1px 1px 0 rgba(0, 0, 0, 0.05);
-    background-color: white;
-    border: 1px solid #d1d5d9;
-    width: 50px;
-    height: 28px;
-    line-height: 28px;
-    font-size: 13px;
-    font-weight: bold;
-    text-align: center;
-    color: #808991;
-  `,
-};
 
 const mainStyles = {
   paddingForPagination: css`
@@ -81,51 +29,7 @@ const mainStyles = {
   `,
 };
 
-function NavigationBar({ shelfTitle, onBackClick }) {
-  return (
-    <Responsive css={navigationBarStyles.wrapper}>
-      <div css={navigationBarStyles.bar}>
-        <button type="button" css={navigationBarStyles.iconWrapper} onClick={onBackClick}>
-          <ArrowLeft css={navigationBarStyles.icon} />
-          <span className="a11y">뒤로 가기</span>
-        </button>
-        <h2 css={navigationBarStyles.title}>‘{shelfTitle}’에 추가</h2>
-      </div>
-    </Responsive>
-  );
-}
-
-function SearchBar({ isSearching, keyword, onClear, onConfirm, onKeywordChange }) {
-  const [isFocused, setFocused] = React.useState(false);
-  const handleFocus = React.useCallback(() => setFocused(true), []);
-  const handleBlur = React.useCallback(() => setFocused(false), []);
-  const handleCancelButtonClick = React.useCallback(
-    () => {
-      onKeywordChange && onKeywordChange('');
-      onClear && onClear();
-    },
-    [onClear, onKeywordChange],
-  );
-  const left = (
-    <SearchBox
-      keyword={keyword}
-      onBlur={handleBlur}
-      onClear={onClear}
-      onFocus={handleFocus}
-      onSubmit={onConfirm}
-      onKeywordChange={onKeywordChange}
-    />
-  );
-  const right =
-    isSearching && !isFocused ? (
-      <button type="button" css={searchBarStyles.cancelButton} onClick={handleCancelButtonClick}>
-        취소
-      </button>
-    ) : null;
-  return <FlexBar css={searchBarStyles.bar} flexLeft={left} flexRight={right} />;
-}
-
-function SearchView({ clearSelectedItems, isSelected, onAddSelected, onBackClick, shelfTitle, uuid }) {
+function SearchModal({ clearSelectedItems, isSelected, onAddSelected, onBackClick, shelfTitle, uuid }) {
   const [keyword, setKeyword] = React.useState('');
   const [isSearching, setSearching] = React.useState(false);
   const [searchItems, setSearchItems] = React.useState(null);
@@ -269,4 +173,4 @@ const mapDispatchToProps = {
 export default connect(
   mapStateToProps,
   mapDispatchToProps,
-)(SearchView);
+)(SearchModal);
